requirements.txt
opencv-python==4.10.0.84
ultralytics==8.2.41
Flask==3.0.3
awscli==1.34.0
boto3==1.35.0
python-dotenv==1.0.1
requests==2.32.3

dill==0.3.8
lapx==0.5.9.post1